+++ src/clj_kondo/impl/config.clj
@@ -30,9 +30,11 @@
               :unreachable-code {:level :warning}
               :datalog-syntax {:level :error}
               :unbound-destructuring-default {:level :warning}
+              :used-underscored-binding {:level :off}
               :unused-binding {:level :warning
                                ;;:exclude-destructured-keys-in-fn-args false
                                ;;:exclude-destructured-as false
+                               ;;:exclude-defmulti-args false
                                ,}
               :unsorted-required-namespaces {:level :off}
               :unused-namespace {:level :warning
@@ -78,8 +80,12 @@
                           :exclude #{}}
               :use {:level :warning}
               :missing-else-branch {:level :warning}
+              :duplicate-case-test-constant {:level :error}
               :type-mismatch {:level :error}
               :missing-docstring {:level :off}
+              :docstring-blank {:level :warning}
+              :docstring-no-summary {:level :off}
+              :docstring-leading-trailing-whitespace {:level :off}
               :consistent-alias {:level :warning
                                  ;; warn when alias for clojure.string is
                                  ;; different from str
@@ -113,6 +119,7 @@
               compojure.core/defroutes clojure.core/def
               compojure.core/let-routes clojure.core/let}
     :macroexpand {}
+    ;; :auto-load-configs true
     :output {:format :text ;; or :edn
              :summary true ;; outputs summary at end, only applicable to output :text
              ;; outputs analyzed var definitions and usages of them